% Generated by build script
\name{hpa_gene}
\alias{hpa_gene}
\title{Fetch Human Protein Atlas gene detail by Ensembl ID}
\usage{
hpa_gene()
}
\arguments{
\item{ensembl_id}{Ensembl gene ID (e.g. "ENSG00000012048")}
\item{columns}{Column codes (default "g,eg,up,gd,chr,pc")}
}
\value{tibble: gene, ensembl, uniprot, gene_description, chromosome,}
\description{Fetch Human Protein Atlas gene detail by Ensembl ID}
